requirements.txt
Django==5.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-environ==0.11.2
django-cors-headers==4.3.1
django-filter==23.5
reportlab==4.0.7
Pillow==11.0.0
mysqlclient==2.2.0
setuptools
django-debug-toolbar